USPS Mail edit
Commented out the subtraction of 1 for the USPS mail. During testing with forwarded past emails the code was counting in the 'attachments.txt' file that is sent. In a real receipt of the informed delivery with a single mail item it was only counting 1 and subtracting 1 to compensate for the 'attachments.txt' was making it report zero mail.