Fix order sensitivity of sensor computation. (#327)
* Fix order sensitivity of sensor computation.

The update() method of the sensor iterates through sensors in the order
they appear in the _resources list. Some sensors, such as the
*_delivering and zpackages_* sensors, depend on the values of earlier
sensors to compute their value. If the _resources list is not in exactly
the right order, this causes these sensors to show the wrong value.
Performing setup through the UI does not sort this list, and can cause
the summary sensors (zpackages_*) to break.

To fix this issue, abstract the "fetch one sensor" logic into a fetch()
method. That method can recursively call itself to fetch the value of
any of its dependencies. The values are cached, so that a given sensor
will only be computed once per update() pass, regardless of how many
times that sensor is fetch()ed.